+++ demos/drawFancyCrosshair.js
@@ -5,11 +5,11 @@
  * @date   2020-11-25
  **/
 
-var drawFancyCrosshair = function( pb, position, isHighlighted, isSelected ) {
-    var color = isSelected ? 'red' : isHighlighted ? 'rgba(192,0,0,0.5)' : 'rgba(0,192,192,0.5)';
-    var lineWidth = isSelected ? 2.0 : 1.0;
-    var crossRadius = 2;
-    var arcRadius = 3;
+var drawFancyCrosshair = function( pb, position, color, lineWidth, radius ) { // isHighlighted, isSelected ) {
+    // var color = isSelected ? 'red' : isHighlighted ? 'rgba(192,0,0,0.5)' : 'rgba(0,192,192,0.5)';
+    var lineWidth = lineWidth || 1.0; // isSelected ? 2.0 : 1.0;
+    var crossRadius = radius * (2/3); // 2;
+    var arcRadius = radius; // 3;
     var s = Math.sin(Math.PI/4)*crossRadius;
     var c = Math.cos(Math.PI/4)*crossRadius;
     pb.draw.line( new Vertex( position.x + c,
@@ -27,6 +27,8 @@ var drawFancyCrosshair = function( pb, position, isHighlighted, isSelected ) {
 			   arcRadius,
 			   Math.PI/2 * (i+1) + Math.PI*2*0.2,
 			   Math.PI/2 * (i+1) + Math.PI*2*0.3,
-			   color, lineWidth );
+			   color,
+			   lineWidth
+			 );
     }
 };